refactor fetch into createprofile component
