+++ src/components/CreateProfile/backend.js
@@ -1,0 +1,3 @@
+const backend = 'http://localhost:3000';
+
+export default backend;
