+++ src/reducers/currentUser.js
@@ -5,6 +5,9 @@ const currentUser = (state = {}, action) => {
   case 'ADD_CITY':
     return action.updatedUser;
   case 'SUBMIT_PROFILE':
+    return action.user;
+  case 'ADD_USER_ID':
+    console.log('in reducer');
     return action.newUser;
   case 'EDIT_PROFILE':
     return action.newUser;
